+++ src/Sentry.Unity.Editor/Native/BuildPostProcess.cs
@@ -4,13 +4,14 @@ using Sentry.Extensibility;
 using UnityEditor;
 using UnityEditor.Build;
 using UnityEditor.Callbacks;
+using System.Diagnostics;
 
 namespace Sentry.Unity.Editor.Native
 {
     public static class BuildPostProcess
     {
         [PostProcessBuild(1)]
-        public static void OnPostProcessBuild(BuildTarget target, string pathToProject)
+        public static void OnPostProcessBuild(BuildTarget target, string executablePath)
         {
             if (target is not (BuildTarget.StandaloneWindows or BuildTarget.StandaloneWindows64))
             {
@@ -18,7 +19,7 @@ namespace Sentry.Unity.Editor.Native
             }
 
             var options = ScriptableSentryUnityOptions.LoadSentryUnityOptions(BuildPipeline.isBuildingPlayer);
-            var logger = options?.DiagnosticLogger ?? new UnityLogger(new SentryUnityOptions());
+            var logger = options?.DiagnosticLogger ?? new UnityLogger(options ?? new SentryUnityOptions());
 
             try
             {
@@ -28,7 +29,7 @@ namespace Sentry.Unity.Editor.Native
                     return;
                 }
 
-                if (options?.Validate() is not true)
+                if (options?.IsValid() is not true)
                 {
                     logger.LogWarning("Failed to validate Sentry Options. Native support disabled.");
                     return;
@@ -40,43 +41,86 @@ namespace Sentry.Unity.Editor.Native
                     return;
                 }
 
-                var crashpadPath = Path.GetFullPath(Path.Combine("Packages", SentryPackageInfo.GetName(), "Plugins",
-                    "Windows", "Sentry", "crashpad_handler.exe"));
-                var targetPath = Path.Combine(Path.GetDirectoryName(pathToProject), Path.GetFileName(crashpadPath));
-                File.Copy(crashpadPath, targetPath, true);
+                var projectDir = Path.GetDirectoryName(executablePath);
+                AddCrashHandler(logger, projectDir);
+                UploadDebugSymbols(logger, projectDir, Path.GetFileName(executablePath));
 
-                // TODO symbol upload
-                // var sentryCliOptions = SentryCliOptions.LoadCliOptions();
-                // if (!sentryCliOptions.UploadSymbols)
-                // {
-                //     logger.LogDebug("Automated symbols upload has been disabled.");
-                //     return;
-                // }
-                //
-                // if (EditorUserBuildSettings.development && !sentryCliOptions.UploadDevelopmentSymbols)`
-                // {
-                //     logger.LogDebug("Automated symbols upload for development builds has been disabled.");
-                //     return;
-                // }
-                //
-                // if (!sentryCliOptions.Validate(logger))
-                // {
-                //     logger.LogWarning("sentry-cli validation failed. Symbols will not be uploaded." +
-                //                        "\nYou can disable this warning by disabling the automated symbols upload under " +
-                //                        "Tools -> Sentry -> Editor");
-                //     return;
-                // }
-                //
-                // SentryCli.CreateSentryProperties(pathToProject, sentryCliOptions);
-                // SentryCli.AddExecutableToXcodeProject(pathToProject, logger);
-                //
-                // sentryXcodeProject.AddBuildPhaseSymbolUpload(logger);
             }
             catch (Exception e)
             {
-                logger.LogError("Failed to add the Sentry crash handler to the built application", e);
+                logger.LogError("Failed to add the Sentry native integration to the built application", e);
                 throw new BuildFailedException("Sentry Native BuildPostProcess failed");
             }
+        }
+
+        private static void AddCrashHandler(IDiagnosticLogger logger, string projectDir)
+        {
+            var crashpadPath = Path.GetFullPath(Path.Combine("Packages", SentryPackageInfo.GetName(), "Plugins",
+                "Windows", "Sentry", "crashpad_handler.exe"));
+            var targetPath = Path.Combine(projectDir, Path.GetFileName(crashpadPath));
+            logger.LogInfo("Copying the native crash handler '{0}' to the output directory", Path.GetFileName(crashpadPath));
+            File.Copy(crashpadPath, targetPath, true);
+        }
+
+        private static void UploadDebugSymbols(IDiagnosticLogger logger, string projectDir, string executableName)
+        {
+            var cliOptions = SentryCliOptions.LoadCliOptions();
+            if (!cliOptions.IsValid(logger))
+            {
+                return;
+            }
+
+            logger.LogInfo("Uploading debugging information using sentry-cli in {0}", projectDir);
+
+            var paths = "";
+            Func<string, bool> addPath = (string name) =>
+            {
+                var fullPath = Path.Combine(projectDir, name);
+                if (Directory.Exists(fullPath) || File.Exists(fullPath))
+                {
+                    paths += " " + name;
+                    logger.LogDebug($"Adding '{name}' to the debug-info upload");
+                    return true;
+                }
+                else
+                {
+                    logger.LogWarning($"Coudn't find '{name}' - debug symbol upload will be incomplete");
+                    return false;
+                }
+            };
+
+            addPath(executableName);
+            addPath("GameAssembly.dll");
+            addPath("UnityPlayer.dll");
+            addPath(Path.GetFileNameWithoutExtension(executableName) + "_BackUpThisFolder_ButDontShipItWithYourGame");
+
+            // no debug info in sentry.dll but sentry.io will complain about a missing file...
+            // addPath(Path.GetFileNameWithoutExtension(executableName) + "_Data/Plugins/x86_64/sentry.dll");
+
+            // Configure the process using the StartInfo properties.
+            var process = new Process
+            {
+                StartInfo = new ProcessStartInfo
+                {
+                    FileName = SentryCli.SetupSentryCli(),
+                    WorkingDirectory = projectDir,
+                    Arguments = "upload-dif " + paths,
+                    UseShellExecute = false,
+                    RedirectStandardOutput = true,
+                    RedirectStandardError = true,
+                    CreateNoWindow = true
+                }
+            };
+
+            process.StartInfo.EnvironmentVariables["SENTRY_ORG"] = cliOptions.Organization;
+            process.StartInfo.EnvironmentVariables["SENTRY_PROJECT"] = cliOptions.Project;
+            process.StartInfo.EnvironmentVariables["SENTRY_AUTH_TOKEN"] = cliOptions.Auth;
+            process.OutputDataReceived += (sender, args) => logger.LogDebug($"sentry-cli: {args.Data.ToString()}");
+            process.ErrorDataReceived += (sender, args) => logger.LogError($"sentry-cli: {args.Data.ToString()}");
+            process.Start();
+            process.BeginOutputReadLine();
+            process.BeginErrorReadLine();
+            process.WaitForExit();
         }
     }
 }